refactor routing and functions
